turn on noImplicitAny
Since this module is designed to be compiled by consumer projects (from what I can tell - see #15), it's nice to turn on noImplicitAny.

The compose function had a signature, but tsc didn't seem to be able to infer the argument types. I've added them, and checked that
`ComposeSignature` is still the type consumers see.

Fixes #15.